frontend / scan: add blindscan support
frontend:
Observe a long timeout for blindscan iterations.

scan:
For blindscan, the initial list of transponders does not need to contain
valid transponders.
Each of the provided transponders will be iterated (i.e. tuned with
blindscan parameter) several times until a tune failure occurs. Each
time a blindscan tune iteration returns ok, a new transponder has been
found and will be scanned.
Each transponder in the initial transponder list causes a full blindscan
iteration run.

Some of the parameters of the initial transponders will be used for the
blindscan, but most will be ignored.

For DVB-S, you need to provide a list of 4 transponders for each orbital
position:
one for each polarity (H/V or L/R), one for each band (hi/lo).
The frequency defines the starting frequency within the desired band.
The symbolrate defines the frequency search range, in MHz (frequency /
1000).
The polarity defines on which polarity the search should run.
All remaining transponder parameters will be ignored.
So for each orbital position, 4 blindscan iteration runs will be done,
one for each polarity/band 'quadrant'.

For DVB-C, only one initial transponder has to be provided.
The frequency defines the start of the blindscan.
The symbolrate defines the frequency search range, in MHz (frequency /
1000000).

For DVB-T, usually only one initial transponder has to be provided.
The frequency defines the start of the blindscan.
The bandwidth defines both the search step as well as the search
bandwidth.

TODO: provide a userinterface for blind scan