Only use blit hardware acceleration for cases where the hardware can do so - thx Peter Urbanec
It is not possible to implement blitAlphaTest easily using hardware
acceleration because the hardware does not have a facility to perform
comparisons on pixel values as part of the blit process. Use hardware
accelerated blits only in the copy and blend modes, perform alpha test
in software.

Make alphaTest mask consistent for all cases and use the correct value - thx Peter Urbanec

The code used a mix of "test high bit" and "test for zero value" for
the alphaTest blit mask. Replace the constant with a symbolic version
and ensure that the test is correctly applied to "zero value" case, not
all cases where the high bit is set.