Introduce fisrt 'tryout' of the Remote Timer integration
Can be enabled via SetupFallbackTuner. At this moment the URL from the
fallbacktuner is used. When the feature is enabled on a timer entry you
have on top an extra option to select if you want to program the timer
locally or remotely. When there are no tuners the default value is True
for remote. When there are tuners enabled in the box the default will be
'local' a timer.

TODO: Most code is derived from the RemoteTimer plugin. In this plugin
not all timer features were supported. E.g. repeated timers and
selecting a directory were not possible. This commit has for now the
same limitations
It must be investigated if OWIF does indeed support all timer features
fist in TimerEntry.py so the timer on the remote (fallback) tuner can be
correctly set/edited. For now the only 'target' that needs a change is
indeed the TimerEntry.py (beside other suggestions).