Second tryout for integrated fallback timer list
It should now be compatible with partnerbox/remote timer
It can also program repeated timers which is not possible with
partnerbox/remote timer. In addition you can also enable/disable timers
on the remote box.

TODO: error handling when the communication with the fallback tuner
'drops' and handling verification if the programming on the remote box
was succesfull and possible other issues you find which I could not find
yet :)